Refactoring Plugin protocol
