+++ SwiftBar/MenuBar/MenuBarItem.swift
@@ -5,9 +5,6 @@ import SwiftUI
 
 class MenubarItem: NSObject {
     var plugin: Plugin?
-    var executablePlugin: ExecutablePlugin? {
-        plugin?.executablePlugin
-    }
 
     private lazy var workQueue: OperationQueue = {
         let providerQueue = OperationQueue()
@@ -85,7 +82,7 @@ class MenubarItem: NSObject {
             barItem.button?.window?.delegate = self
         }
         updateMenu()
-        contentUpdateCancellable = executablePlugin?.contentUpdatePublisher
+        contentUpdateCancellable = plugin?.contentUpdatePublisher
             .receive(on: menuUpdateQueue)
             .sink { [weak self] _ in
                 guard self?.isOpen == false else {
@@ -134,7 +131,7 @@ extension MenubarItem: NSMenuDelegate {
         params.params["color"] = "white"
         barItem.button?.attributedTitle = atributedTitle(with: params).title
 
-        guard let lastUpdated = executablePlugin?.lastUpdated else { return }
+        guard let lastUpdated = plugin?.lastUpdated else { return }
         let formatter = RelativeDateTimeFormatter()
         formatter.unitsStyle = .full
         let relativeDate = formatter.localizedString(for: lastUpdated, relativeTo: Date()).capitalized
@@ -347,7 +344,7 @@ extension MenubarItem {
         statusBarMenu.removeAllItems()
         show()
 
-        if executablePlugin?.lastState == .Failed {
+        if plugin?.lastState == .Failed {
             titleLines = ["⚠️"]
             barItem.button?.title = "⚠️"
             buildStandardMenu()
