+++ SwiftBar/Plugin/Plugin.swift
@@ -1,3 +1,4 @@
+import Combine
 import Foundation
 import ShellOut
 
@@ -15,28 +16,28 @@ enum PluginState {
 
 typealias PluginID = String
 
-protocol Plugin {
+protocol Plugin: AnyObject {
     var id: PluginID { get }
     var type: PluginType { get }
     var name: String { get }
     var file: String { get }
     var enabled: Bool { get }
-    var metadata: PluginMetadata? { get }
+    var metadata: PluginMetadata? { get set }
+    var contentUpdatePublisher: PassthroughSubject<Any, Never> { get set }
     var updateInterval: Double { get }
     var lastUpdated: Date? { get set }
     var lastState: PluginState { get set }
     var content: String? { get set }
     var error: ShellOutError? { get set }
     func refresh()
-    func terminate()
-    func invoke(params: [String]) -> String?
+    func enable()
+    func disable()
+    func invoke() -> String?
+    func makeScriptExecutable(file: String)
+    func refreshPluginMetadata()
 }
 
 extension Plugin {
-    var executablePlugin: ExecutablePlugin? {
-        self as? ExecutablePlugin
-    }
-
     var description: String {
         """
         id: \(id)
@@ -53,5 +54,44 @@ extension Plugin {
             dependencies: \(metadata?.dependencies?.joined(separator: ",") ?? "")
             aboutURL: \(metadata?.aboutURL?.absoluteString ?? "")
         """
+    }
+
+    var prefs: Preferences {
+        Preferences.shared
+    }
+
+    var enabled: Bool {
+        !prefs.disabledPlugins.contains(id)
+    }
+
+    func disable() {
+        lastState = .Disabled
+        prefs.disabledPlugins.append(id)
+    }
+
+    func enable() {
+        prefs.disabledPlugins.removeAll(where: { $0 == id })
+        refresh()
+    }
+
+    func makeScriptExecutable(file: String) {
+        guard prefs.makePluginExecutable else { return }
+        let script = """
+        if [[ -x "\(file)" ]]
+        then
+            echo "File "\(file)" is executable"
+        else
+            chmod +x "\(file)"
+        fi
+        """
+        _ = try? shellOut(to: script)
+    }
+
+    func refreshPluginMetadata() {
+        let url = URL(fileURLWithPath: file)
+        metadata = PluginMetadata.parser(fileURL: url)
+        if let script = try? String(contentsOf: url) {
+            metadata = PluginMetadata.parser(script: script)
+        }
     }
 }
